+++ src/main/kotlin/com/lykke/matching/engine/database/azure/AzureDictionariesDatabaseAccessor.kt
@@ -3,11 +3,13 @@ package com.lykke.matching.engine.database.azure
 import com.lykke.matching.engine.daos.AssetPair
 import com.lykke.matching.engine.daos.azure.AzureAssetPair
 import com.lykke.matching.engine.database.DictionariesDatabaseAccessor
+import com.lykke.matching.engine.utils.RoundingUtils
 import com.lykke.utils.logging.MetricsLogger
 import com.lykke.utils.logging.ThrottlingLogger
 import com.microsoft.azure.storage.table.CloudTable
 import com.microsoft.azure.storage.table.TableOperation
 import com.microsoft.azure.storage.table.TableQuery
+import java.math.BigDecimal
 import java.util.HashMap
 
 class AzureDictionariesDatabaseAccessor(dictsConfig: String): DictionariesDatabaseAccessor {
@@ -28,7 +30,12 @@ class AzureDictionariesDatabaseAccessor(dictsConfig: String): DictionariesDataba
                     .where(TableQuery.generateFilterCondition("PartitionKey", TableQuery.QueryComparisons.EQUAL, ASSET_PAIR))
 
             for (asset in assetsTable.execute(partitionQuery)) {
-                result[asset.assetPairId] = AssetPair(asset.assetPairId, asset.baseAssetId, asset.quotingAssetId, asset.accuracy, asset.minVolume, asset.minInvertedVolume)
+                result[asset.assetPairId] = AssetPair(asset.assetPairId,
+                        asset.baseAssetId,
+                        asset.quotingAssetId,
+                        asset.accuracy,
+                        BigDecimal.valueOf(asset.minVolume),
+                        BigDecimal.valueOf(asset.minInvertedVolume))
             }
         } catch(e: Exception) {
             LOGGER.error("Unable to load asset pairs", e)
@@ -43,7 +50,7 @@ class AzureDictionariesDatabaseAccessor(dictsConfig: String): DictionariesDataba
             val retrieveAssetPair = TableOperation.retrieve(ASSET_PAIR, assetId, AzureAssetPair::class.java)
             val assetPair = assetsTable.execute(retrieveAssetPair).getResultAsType<AzureAssetPair>()
             if (assetPair != null) {
-                return AssetPair(assetPair.assetPairId, assetPair.baseAssetId, assetPair.quotingAssetId, assetPair.accuracy, assetPair.minVolume, assetPair.minInvertedVolume)
+                return AssetPair(assetPair.assetPairId, assetPair.baseAssetId, assetPair.quotingAssetId, assetPair.accuracy, BigDecimal.valueOf(assetPair.minVolume), BigDecimal.valueOf(assetPair.minInvertedVolume))
             }
         } catch(e: Exception) {
             if (throwException) {
